Minor formatting and some casts
Former-commit-id: 5493889b5c05bb608e06dcf5316a48e46686c4a5 [formerly 02fd0e7c4597ab0069e3934a64f477da3ccc760b]
Former-commit-id: 49b1975903d1ac47a67010ce3342de01b3e675d6

+++ safecache.cpp
@@ -45,7 +45,7 @@ bool SafeCache::load(const char *filename, int length)
     if (fstat(fd, &statbuf) == 0) {
         if (statbuf.st_size > 0 && (length < 0 || length > statbuf.st_size)) {
             if (statbuf.st_size > capacity()) length = capacity();
-            else length = statbuf.st_size;
+            else length = static_cast<int>(statbuf.st_size);
         }
     }
     if (length >= 0) resize(length);
@@ -54,7 +54,7 @@ bool SafeCache::load(const char *filename, int length)
     length = size();
     int remain = length;
     while (remain > 0) {
-        ssize_t chunk = read(fd, data() + (length - remain), remain);
+        ssize_t chunk = read(fd, data() + (length - remain), static_cast<size_t>(remain));
         if (chunk < 0) goto ending;
         remain -= chunk;
         fsync(fd);
@@ -69,9 +69,9 @@ bool SafeCache::load(const char *filename, int length)
 bool SafeCache::save(const char *filename, mode_t mode)
 {
     bool ok = false;
-    int fd = open(filename, O_CREAT|O_TRUNC|O_WRONLY, mode);
+    int fd = open(filename, O_CREAT | O_TRUNC | O_WRONLY, mode);
     if (fd == -1) return false;
-    if (write(fd, constData(), size()) == size()) goto ending;
+    if (write(fd, constData(), static_cast<size_t>(size())) == size()) goto ending;
     if (fchmod(fd, mode) != 0) goto ending;
     ok = true;
